requirements.txt
asgiref==3.9.1
Django==5.2.4
pillow==11.3.0
PyMySQL==1.1.2
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers
gunicorn
google-genai